+++ src/adapters/profile.ts
@@ -105,7 +105,7 @@ class Profile extends Fetcher {
     return await res?.json()
   }
 
-  public async getUserActivities(profileId: string, page = 0, size = 8) {
+  public async getUserActivities(profileId: string, page = 0, size = 12) {
     return await this.jsonFetch(
       `${MOCHI_PROFILE_API_BASE_URL}/profiles/${profileId}/activities?page=${page}&size=${size}`
     )
